requirements.txt
requests
dotenv
os
discord.py